nanodet-main/requirements.txt
Cython
termcolor
numpy
torchvision
tensorboard

matplotlib
pyaml
opencv-python
tqdm